requirements.txt
streamlit==1.42.1 
python-dotenv==1.0.1
google-generativeai==0.8.4
pymupdf==1.25.3